Функция get_new_posts перенесена в класс AutoPoster Изменена структура файлов
