+++ TG_AutoPoster/__init__.py
@@ -1,5 +1,4 @@
 from .TG_AutoPoster import AutoPoster
-from .utils.vk import get_new_posts
 
 __author__ = "Adrian Polyakov (me@qwertyadrian.ru)"
 __version__ = "3.0.0"
