+++ TG_AutoPoster/utils/__init__.py
@@ -1,0 +1,5 @@
+from .vk import *
+from .tg import *
+
+from TG_AutoPoster.utils.vk.sender import Sender
+from .tools import *
